+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Added
 - **Image Upload / Manage**: Allow reupload not found images, uses same url/id
+- **Editor Toolbar**: impl empty spaces, lines, Karo-Muster for Work Tasks
 
 
 ## [0.6.7] - 2026-06-25